// src/components/types/book.ts
export enum PaginationBtn {
  START = 'dbl-prev',
  PREV = 'prev',
  NEXT = 'next',
  FINISH = 'dbl-next',
}

export enum PageValue {
  ONE = 1,
  TWO = 2,
  THREE = 3,
  FOUR = 4,
  FIVE = 5,
  SIX = 6,
  SEVEN = 7,
}

export enum PathBackgroundBook {
  BACKGROUND_1 = './background-book-1.svg',
  BACKGROUND_2 = './background-book-2.svg',
  BACKGROUND_3 = './background-book-3.svg',
  BACKGROUND_4 = './background-book-4.svg',
  BACKGROUND_5 = './background-book-5.svg',
  BACKGROUND_6 = './background-book-6.svg',
}